Fix AMBR lookup to handle unmatched type name
The first pass implementation of the AMBR type lookup used QuerySet.first()
after filtering to names that match exactly or are matched via
MeasurementNameTransform mappings. However, QuerySet.first() will happily
return `None` when no match is found, where we want an error. Switching to a
call to QuerySet.get() instead gets the behavior we want, but also requires
handling the case where multiple types match a name.

While doing this, also tighten up the looping code, so that each column pair is
iterated once to generate a 2-tuple of values, instead of twice to generate two
lists that are ultimately passed to `zip()` to make a 2-tuple. Pulling out the
"decimation" sampling into its own function, to make it clearer that we're
passing through a sample filter before mapping to EDD import tuples.

Also making sure to add tests to cover all of the AMBR parser functions.