Add guaranteed unique Assay names in test factory
On occaision, some tests of the edd.load import code would fail, due to the
randomly generated Assay names in the test code just happening to chose the
same name for sequential calls. There is a small, but very much non-zero,
chance of this happening, and it will randomly fail around 5% of test runs.
Switching the `name` field to a lazy attribute that incorporates the parent
line name and a sequence number will guarantee the uniqueness, and prevent
these random test failures.

Additional changes are coming from a newer `black` version formatting the code.
The added line breaks are small enough a change to keep without making a
separate commit.

Fixes: EDD-1436