Update node packages for frontend build
Several node packages are outdated, and it's started to cause problems with
builds. As of a few weeks ago, builds would give a warning message with a
mismatch between the ESLint packages and the TypeScript packages. All these are
updated. A build today, however, started failing with this message:

```
Error: Debug Failure. False expression: Non-string value passed to
`ts.resolveTypeReferenceDirective`, likely by a wrapping package working with
an outdated `resolveTypeReferenceDirectives` signature. This is probably not a
problem in TS itself.
```

This error is due to a breaking change in an TypeScript compiler API, as
interacted with via the typescript loader (I think). Thus, the node packages
are due for a refresh. Most changes are simple drop-in.

The main exceptions are the upgrade to Webpack 5, and an update to the APIs for
`react-dropzone`. The Webpack upgrades replace a few settings with new names,
and uses the new Asset Module for fonts, instead of `url-loader`/`file-loader`.
The `react-dropzone` API changed a minor interface type, requiring an update to
its uses in the Wizard import for Step 2.

Updating the Webpack settings results in a minor change to the styles for the
LoadWizard entrypoint. The plugin handling CSS was configured with a
`chunkFilename` setting of `[name].css` and a `filename` of `styles.css`. It is
now using a consistent `[name].css` for both. Switching the template to load
`dist/LoadWizard.css` instead of `dist/styles.css` now does the correct thing.