Remove doc references to Docker Hub images
The images for EDD pushed to Docker Hub are not currently kept up-to-date, and
lag significantly behind the production code. Mixing current code with the
outdated images causes lots of problems. Our deployment process no longer
relies on Docker Hub, and there's no bandwidth to support keeping those images
current. To prevent this from being a problem, update the documentation to
remove references to Docker Hub image versions, and only support
built-it-yourself images.