Add a display name field to user profiles
This change adds a field to every user's Profile for a display name. This is
overall good practice to have, given reasons listed in "Falsehoods Programmers
Believe About Names" ([link][names]), and that Django's defaults violate this
by having `first_name` and `last_name` fields. Adding a field for this allows
for a consistent and *updatable* bit of text to use to refer to a specific
person or account.

Additionally, adding this field enables an easier implementation of a nifty
feature, where a single autocomplete field can be used to select a user *or*
group for things like setting permissions. With a properly constructed UNION
query, we can search through both kinds of records and deal with the response
records in a single query. Without it, two different queries would need to
happen, then merge the results, then order the merged results in application
code … and it's a huge mess. Implementing convoluted logic to "compute" this
value for users from multiple different fields is prone to error, as well.

[names]: https://www.kalzumeus.com/2010/06/17/falsehoods-programmers-believe-about-names/