Update build with split Bootstrap v3 and v5
We are migrating from Bootstrap v3 to Bootstrap v5. This is a major shift in
the frontend code, and doing a hard cutover would be a huge change to do all at
once. Instead, update the build to have two pipelines to create the assets on
the frontend with each version. This way, templates can update in smaller
chunks that are easier to reason about.

This first step builds the common script to load into every page of the
Bootstrap v5 interface, and the script for login pages. All the login-related
pages here are using the updated Bootstrap v5 interface, and all other pages in
EDD can continue using the current interface that uses parts of Bootstrap v3.

Part of the switch to Bootstrap v5 is dropping JQuery UI, and the abandoned
autocomplete plugin we have that extends JQuery UI. Instead, we're using
Select2, which is also the library used by Django in the default admin
interface. A few backend changes to support the new autocomplete APIs are added
here, along with the frontend changes.