Update index page with Bootstrap v5 interface
A big piece of making the index page more accessible was replacing the
"homegrown" DataGrid table library with a third-party table library that has
better a11y. The index page is now using the `datatables.net` library, and the
rest of the page is updated to follow the interface conventions of Bootstrap
v5. This includes some minor tweaks to the backend providing data on Study
objects for the table to display, with additional information used to render
the table and paging properly.