Fix test warning about unclosed file
The test for deleting an attachment would generate warnings about an unclosed
file. Deleting an Attachment model would remove the record from the database,
but not the file; and it appears that in some cases, manipulating the model
object will open the file. Adding an explicit delete step to the model delete,
so the underlying file should be removed before the database record is removed.
Also adding an explicit `open()` context manager when retreiving attachments.

The warning fixed by this change looks like this:
```
main/tests/test_views.py::StudyAttachmentViewTests::test_post_after_confirmation
  /usr/local/lib/python3.10/site-packages/django/db/models/sql/query.py:1532: ResourceWarning: unclosed file <_io.FileIO name='/var/www/uploads/2022/07/27/ImportData_FBA_HPLC_f8HjOdW.xlsx' mode='rb' closefd=True>
    child_clause, needed_inner = self.build_filter(
```