Format main urls.py path definitions
Some upcoming changes will modify the paths used for operations on Study
instances. As some path definitions will fit in a single line, while others
will be forced into a multi-line layout due to length, making a formatting
change here to pre-emptively make all the paths multi-line, so that it's quick
and easy to pick out the paths used.