Fix "null" value in old autocomplete
If the old autocomplete has a value field with the string "null", or any other
unknown primary key value, it will fail to resolve to an object and raise an
exception. Updating the code to instead check if any object can resolve, and
otherwise treat the value as empty.

Additionally, the frontend code was incorrectly inserting a "null" string in
for a value, which causes the backend form to fail validation. If no contact is
found, instead use an empty value.