Fix false migration detections
This is probably a result of a Django upgrade at some point. During startup of
EDD with existing databases, there will be a log message that looks like this:

    Your models in app(s): 'campaign', 'main' have changes that are not yet reflected in a migration, and so won't be applied.
    Run 'manage.py makemigrations' to make new migrations, and then re-run 'manage.py migrate' to apply them.

Creating the migrations will add a migration each in `edd/campain/migrations`
and `main/migrations`, to alter the `updates` field with a different
`related_name`. This change results in a SQL noop if running through `manage.py
sqlmigrate`. None of the code for this field has changed, so it is safe to
retroactively modify the migration that adds the fields with the correct
`related_name`. That eliminates the detected changes.