Remove repeated Dockerfile stage
The Dockerfile used a `TARGET` argument to label parent images in some build
stages, with repeated sections of the Dockerfile allowing for switching between
slight variations (having dev packages installed, or not). Functionality for
doing this switching is better off in a simple shell conditional, without
repeating large parts of the Dockerfile.