Add error check to EDD startup
Found this while working on setting up EDD to use `swarmDeploy()` Jenkins
function. If no settings are passed in, the startup code check to copy settings
into the runtime directory would have a non-zero exit. This causes EDD to exit
entirely, with no indication of what is wrong. Now, it will print a message
noting that no settings are provided, and continue normally.