Fix changing permission for an entity
Changing permissions for a user/group or everyone, from Read to Write, or Write
to Read, was using an incorrectly written update call. The result would be two
permission objects, one with Read and one with Write. Updating the call to
correctly update any existing permission and set to the correct access level,
or create a new permission if no existing permission is found.