Fix use of FormData in AJAX requests
A prior iteration of the code was using jQuery's serialization of form data
when passing to `$.ajax()`. However, this does not work with file uploads;
while using the `FormData` DOM object will work with uploads. I switched the
attachment upload to use the additional parameters instructing jQuery to not
try to serialize the object, but missed doing so on the other calls. As a
result, the click handlers would get an exception, and stop processing.