Update Jenkinsfile for shared library calls
Replace all of the custom shell scripts and Groovy code for the EDD build,
test, and deploy pipeline; and instead replace it with the standardized shared
library functions used by other projects. This makes for a much, MUCH simpler
and shorter Jenkinsfile.

The main workflow runs builds for the EDD application, documentation, and the
TypeScript build, followed by unit tests (for now, only the minimal TypeScript
unit tests). If that passes, then an integration environment is setup, and the
full EDD test suite is run.

If all the above completes successfully, then if the pushed images are for
`trunk`, `dev1`, or `dev2`, another deploy runs into those staging envs.