Add default open state for overview permissions
Request from Chris Petzold; since it's very important to correctly set
permissions for collaborators on a study, especially on the ABF instances where
Studies are default-closed, vs the JBEI instances where it's assumed all have
default-read for JBEI members. Add the `open` attribute to the section with
permissions table and update controls, so they start as visible on the page.