Update plot label formatting
Our plot labels were using format of ".2s", which uses "SI prefix" on numbers
it formats. A question came in from ANL about a label showing "100.00m" just
above a value of 0.08319 for OD600. This "m" suffix was confusing, and could be
even more confusing with units on the axis like "mL".

Switching the label format to ".2g" will instead format the label as "0.10",
and in general use exponent notation for very large and very small numbers,
instead of the SI abbreviations. See the docs for this change at
https://github.com/d3/d3-format#locale_format