Add standard non-specific protocols to EDD
Discussions from the team have led to the conclusion that a common list of
non-specific protocols should be provided in the "import wizard" interface. The
idea is, users may quickly indicate the non-specific protocol used from a
narrowed down list, without needing to search through all protocols, nor go
through the process of defining the entire protocol on `protocols.io`. A future
update should extend the wizard interface to allow selecting one of these
non-specific protocols, OR give a protocols.io link.

Adding `__str__` methods to the Load model classes that did not already have
them. Without these, the admin pages would refer to objects as
`Classname (id)`, which is not very informative.