Fix delete confirm page for Lines/Assays
The delete confirmation page was updated with the Study delete action on the
overview page, but broke the confirmation of deletes for Lines and Assays that
are using the same template. Updating those views to properly use the template.
As the stand-alone Delete view is no longer used, removing it in favor of the
"parent" view methods.