Fix autofocus on search field with autocompletes
There is an unintended interaction between JQuery's handling of focus events
and the way Select2 works, where with specific versions of each library, the
search field part of an autocomplete will fail to focus. To work around the
problem, we can hook into the event where an autocomplete is opened, and use
that to trigger a native browser API focus event, skipping over JQuery's
handling of events.

Presumably, the libraries will update to fix this problem eventually. The
additional event handler and the forceFocus function should be removed once the
fix is merged upstream.

See: https://github.com/select2/select2/issues/5993