Add unique task queues per task
Due to using only one default queue for all tasks, we experience big
performance issues in EDD when that queue gets flooded with tasks that take a
long time. A particular problem appears when several tasks to check UniProt are
in the queue, since the task is rate-limited. With only one queue, if several
of these are in line, nothing else can execute until the backlog is cleared.

By switching to a queue per task, the worker processes can consume messages of
every task. This will let tasks execute, even if a particular queue gets filled
up and is executing slowly.

Using the `task_create_missing_queues` setting to automatically generate a
queue from names, and assigning all current task names to a queue of the same
name. The workers are already starting with only an exchange name, and should
consume messages from all the queues in the exchange.

Note: the line lengths for the task names are long, and flake8 complains about
them. If trying to break the line, black will revert it back. There may be a
better way to do this someday, but not today.

Co-authored-by: William Morrell <WCMorrell@lbl.gov>
Fixes: EDD-1443
Fixes: EDD-1442
Fixes: EDD-1400