Fix API listing measurement types
The previous API was a little convoluted, and excluded information specific to
a kind of measurement unless querying specifically for that kind. Can delete
chunks of code by simply querying for those bits of information, and having the
serializer include them if present.

Making this change will fix a problem where some provisional UniProt proteins
will display as an empty string, because the query for the type information is
not including the accession ID.