Refactor autocomplete search widgets
Adding some better structure for the Metadata autocomplete widgets, so that
uses of the widget can separately configure what kinds of metadata, and whether
to exclude or only show builtin field "metadata". Include a default sort on
type name for metadata searches. Fix some syntax errors in the widgets for SBML
autocompletes, which are not yet converted for use under Bootstrap 5.