Remove custom docker service images
The custom images for RabbitMQ and Redis are only adding health check scripts.
These are not needed in development, so are removed.

Our production deployments are using similar health checks, loaded as config
files. The main repo does not need to decide on the type and frequency of
health checks.

We are no longer using the custom base image for scikit-learn. We're instead
using multi-stage builds, so all the install steps in the previous image are
now included as a build layer.

Other projects that made use of the scikit-learn base can continue to use the
version that is built weekly from the Dockerfile in the internal swarm_config
repo. Those projects should eventually update to include their own build layer.

Moving the Solr schema configs to their own directory, for any development
installs that are using Solr. There will no longer be a custom built Solr
image, as it's more flexible to use the upstream official image and load in our
config files separately. This brings us another step closer to eliminating the
docker subdirectory, and making that part of setup a bit less complicated.