Add utility code to Bootstrap 5 script environment
These files are migrating from similar files used in the previous generation
TypeScript code. They are not used yet as of this commit, but should be in
place for future stages of the Bootstrap 5 conversion project.

The `js.ts` module contains some functions to simplify some checks and
comparisons that "should" be simple one-liners in code, but aren't due to how
JavaScript works. The contents roughly correspond to the non-time functions in
the `JS` class of `typescript/modules/Utl.ts`.

The `access.ts` module is a collection of classes and functionality dealing
with progressively accessing data from the EDD backend, replacing code that
used to try downloading all the data at once before processing on the
front-end. The contents are similar to `typescript/modules/table/Access.ts`,
with minor edits.