Convert study description page to Bootstrap 5
This is a big change. The previous version of the page used a giant single
Django view for all functionality, which made converting it to newer Bootstrap
5 layout a bit difficult. The functionality is now devolved into multiple views
for each action; e.g. `CreateLineView`, `AddAssayView`, `StudyDescriptionView`,
each getting their own URL. Lots of form rendering is now done directly in the
views, instead of relying on front-end scripts to build the forms.

To help make the transition smoother, the new study description page includes a
menu item in its options dropdown, to switch to the "Legacy Table". All the
existing page and functionality remains, if switching to the previous view.
When we're more sure of the updated view's readiness, the old view, scripts,
template, and tests can be deleted. These are marked with a `# DEPRECATED`
comment on relevant sections.