Add parameterized tests of admin pages
The admin site is one of the sections of EDD that has the largest number of
missing coverage lines in tests. Adding in a test file that loads the add or
changelist views of every model in the admin site, and verifies that they load
correctly. Increases total test coverage by nearly 100 lines.

Include a `conftest.py` in the main module tests, to ensure the request
threadlocal variable is cleared before/after every pytest test. Tests using
the Django test classes via `edd.TestCase` are already getting this, but it
needs to happen for functional tests as well as the class-based tests.