Move Dockerfile to project root
This eliminates the last of the items in the `docker` subdirectory. All the
containers dealing directly with EDD code are now in a single multi-stage
`Dockerfile`, and it is stored at the project root. The Nginx containers and
docs have long since migrated to another repo, and are no longer built here.
The code and docs are now built from the project root, instead of needing to
drill down into `docker` directory structure, and everything needed is now
visible immediately to viewers of the repository.

The container orchestration setup is now moved to a `setup` directory. The
`init-config` script is merged into `bin/dev_quickstart.sh`, which should
make getting started with a dev environment much faster.

Scripts to run inside the EDD container(s) are moved to `container-bin`, which
should hopefully have better visibility and be distinct from the local,
non-Docker, `bin` scripts.

Various documentation files within `docker` are now in the `docs` folder.

Container startup scripts are moved to the root, using `entrypoint.sh`,
`invoke.yaml`, and `startup-tasks` directory.