Update dependency management to use Poetry
Many of our other projects are using Poetry for managing Python dependencies,
and EDD is the only active project still using `pipenv`. In the interest of
keeping things consistent, switching over to use Poetry instead. This includes
moving the items previously in `Pipfile` to `pyproject.toml`, and keeping track
of dependencies in `poetry.lock`, as well as keeping separate install manifests
in `requirements.txt`, `requirements.dev.txt` and `mkdocs.requirements.txt` for
easier use in `Dockerfile` builds.

Adding a new script, `dev_lock.sh` for updating all the dependency manifest
files. The typical task of updating dependencies to the latest versions just
requires running the script. Any deviations from the typical task can use the
script as a guideline.

Pinning `openpyxl` to before 3.1 release, as a deprecated function removed in
the 3.1 release is used in the EDD code. Also pinning `django-environ`, as the
0.10.0 release defaults to using Django's new built-in Redis backend, and some
APIs are different vs the `django-redis` backend. These versions will be
un-pinned once the code is updated to support the newer versions.

Fixes: EDD-1459