Add auto-provision admin account to dev
Doing this makes it easier to get started with a pre-provisioned admin
user in the development environment. Before this, the initial user would
have to be manually added to the `http` container.

The account is created using the current user's *nix username and `git
config` email. These are copied into the `http` container via adding
enviornment to the generated `docker-compose.override.yml` file.