Update profile page
Preparing the profile page to include a listing of user preferences and a way
to edit them. There is an "infobox" message currently displayed at the top of
every study description page, indicating that the table of lines is also a
dropzone for experiment setup files. There are also many other pages in EDD
that could usefully have a similar "infobox" describing how to use that page,
but would be cumbersome to keep around all the time. If we add a preference
flag for these "infobox" messages, and a way to dismiss them that flips the
display of them, then users can see the message once, and indicate to not show
it again. We can then allow for seeing all the currently enabled/disabled
settings items on the user's profile page, or linked from the profile page.