Add "server" pyproject.toml
Consolidating configuration for tools into a `pyproject.toml` file. This file
replaces the separate `.coveragerc` and `pytest.ini` files, and is kept
distinct from the root `pyproject.toml` that manages Poetry configuration for
dependencies of the "server" code, as well as the documentation generator,
setup container, and startup/entrypoint invoke script.