Upgrade to Django 4.2
The Django 4.0.x series is officially end-of-life, so it's time to upgrade to
the 4.2.x LTS series. This involves upgrading a handful of other packages,
removing some others, and some minor code changes to take care of warnings.
The packages in `pyproject.toml` are also using more constrained version
ranges, as Poetry takes an excessive amount of time to solve versions if they
are all un-bounded `*` versions. Code changing because of Django warnings:

- Explicitly adding the div form template to ExportSelectionForm, as the
  default form template is deprecated.
- Changing extension of UserAdmin actions list, as the Django class switched
  from defining it as a list to as a tuple.
- Add explicit ordering to Metadata REST API, to fix warnings about it not
  having a defined default ordering.
- Remove definition of the USE_L10N setting; the value is the same as the
  default, and the setting is slated for removal in Django 5.0.
- Skip staticfiles hashing of missing referenced files. The webpack build
  of frontend files includes sourcemap comments for individual components,
  but does not collect those sourcemap files. Django throws an exception
  when not finding the referenced file, but we can simply leave the
  comment as-is.
- Fix the lookup of studies using a MeasurementType in the admin, there is
  a Django warning about using a direct object lookup instead of the
  primary key.
- Update migrations to remove references to the index_together attribute of
  model classes, slated for removal in Django 5.0; replacing the code with
  the indexes attribute, defining an Index on the same fields.
- Fix a test, where previous Django would encode response contents with a
  &quot; HTML entity, it now directly returns a quote character.

The Celery package has just released its 5.3.0 version, allowing for removing
of the pin on the 5.3.0b1 release and fixing several deprecation warnings
coming from Celery / kombu. All existing tests are passing under the
new release.

There is now support for the new psycopg3 library supporting async in Django,
so replacing the psycopg2 package with psycopg version 3.

On the testing side, we were not making use of the pytest-parallel package, and
it is currently unmaintained. With the update, it's raising a handful of
warnings, so pre-emptively removing it before it causes the test runner to
begin failing.