Update build-time collectstatic command
The image build runs the Django collectstatic command to get all the static
files from Django apps into one location. Then, it's much simpler to look in
just that one image location and add in the dist folder from the webpack build
to get the complete set of static assets. We do not need to build a manifest at
the build-time collectstatic, nor do we need to post-process files to add in
file hashes. Updating the code this way should slim down the image by halving
the storage needed for static assets.