Add testing-specific Redis cache configs
For some time, there has been an issue with intermittent test failures when
running EDD's tests, particularly when using a single test environment with
repeated test runs in succession (i.e. development). I had assumed that this
may be a problem with unintended cache collisions, where a randomized username
gets data from a previous test run, where it was expecting a clean environment.
Anecdotally, running the tests again generally fixed the issue, and flushing
out the keys in Redis would stop the issue from appearing for a while.

This config change adds a new file to `edd.settings`, which imports everything
from `edd.settings` and alters the `CACHES` setting to use an alternative cache
URL if it's provided. The pytest config then uses this settings file as it's
Django Settings, so the alternative URL is used for all tests. We have to do it
this way, instead of simply overriding the settings on each test, because the
`django_redis` package fixes its settings at startup time, and overriding
settings after the fact will have no effect.

Also added are two pytest fixtures, one that will run the Redis command
`BGREWRITEAOF` to re-write the append-only file after all tests in the session
are run, and the other to flush (i.e. delete everything) the cache before each
test run. This ensures that every test is guaranteed a clean cache. Both
fixtures only apply when a testing cache is configured.

Enable use of the testing cache with the `CACHE_TEST_URL` environment on the
EDD container. In my development environment, this is the same value as
`CACHE_URL`, with the database number set to `2` instead of `1`, so:
`rediscache://redis:6379/2`.

After making the changes, the cache for regular development use is unaffected
by running tests, and the test cache is always cleaned up and kept small.