Refactor exception reporting in edd.load
Cleaning up some typing annotations dealing with exceptions and reporting in
EDD's import data processing. Annotations using the old `Optional[]` and
`Union[]` typing classes are now using a pipe operator for unions. Adding an
exclude rule to coverage to ignore `if TYPE_CHECKING` blocks.

Switching to structural pattern matching (i.e. the `match` statement) to deal
with handling the `details` parameter to messages. This should resolve bugs
where a lazy-evaluated translation with Django's `gettext` is not recognized as
a string. Where before an exception like `SomeException(details=_("translate
this string"))` would have a details entry of a list of every character in the
translation, now it should only be a single-item list with the translation.

Fixing a few exception types such that the processing to create error messages
happens inside the custom exception, instead of inside each except clause in
the calling code.