Fix slow tests on notifications
Several of the notification tests were annotated with `django_db` and the
`transaction=True` argument. These tests turned out to take an excessive amount
of time with the database, taking around 16-17 seconds each on slower test VMs.
However, removing the annotation from the tests causes the tests to fail. What
appeared to be happening is that the test setup created a fake user in a
transaction, but the async code used a different database connection object to
read back the user ID that's "logged in", so the session information added as a
session cookie does not validate. The fake user ends up not existing, and all
the tests run as if for an anonymous user. The `transaction=True` argument
seems to set a savepoint, create and commit the database record, then instruct
the database to rollback to the savepoint at the end, taking a significant
amount of time.

The `pytest` configuration is now updated to report the ten slowest test calls,
setups, and teardowns. In CI, one of these should always be the test setup that
creates the test database from migrations, currently a process taking about 25
seconds. In development, this can be greatly reduced with the use of the
`--reuse-db` flag to `pytest`, with the caveat that `--create-db` must also be
used whenever any schema changes are made to the code.