Fix some worklist export functions
An EDD worklist export makes use of a mini format language to select properties
of EDD objects, based on the JSON output of those objects. Some tests were
sometimes failing when using data from a Line or Protocol. The Line JSON was
including a `strain_ids` field that doesn't match up with anything on the Line
model. This field is now removed. Protocol was previously making use of the
parent EDDObject function, but it no longer inherits from EDDObject; so an
appropriate JSON representation is added.