Fix tests to never run with Django DEBUG enabled
The DEBUG Django setting can cause some performance penalties when running
code. Update the test runner to set the `EDD_DEBUG` environment, which by
default toggles the Django setting in EDD, to "false". Additionally, add a test
that verifies that tests are run with DEBUG disabled, and additionally that the
Django Debug Toolbar is not included in the application stack.