Remove old squashed migrations
All the migrations on the `main` app were squashed in 2020, greatly increasing
the speed of database setup and test runs. We kept the old migrations around in
case anyone deployed EDD using the old migrations. At this point, it should be
reasonable for anyone using EDD to have upgraded to the point where the
migrations were squashed; and if they have not, they are better off just
starting with a fresh database using the current code.