Update tests to clean up uploaded attachments
The test suite has a few tests that upload attachments as part of the test.
These files would get written to the storage backend configured for the
environment, and could pile up over time. These tests should now be fixed to
clean up after themselves, so there won't be several copies of the same fixture
files duplicated every time a dev environment runs tests.

Part of the update is changing how attachments are handled. They now will
special-case attachments created by the `system` user (typically only from
migrations / bootstrapping code), and store these at the media root. Normal
attachments go into directories for the current date as before, but are now
also prefixed with the parent object ID. This will prevent potential issues
where two users upload a file with the same name, on the same day, but with
different content. With how the code was working, this would result in the
second uploaded file overwriting the first. Now, both will get unique names
based on where the attachment is added. One caveat, if a file is uploaded to
the same study, with the same name, on the same day, the same behavior will
occur. This is obscure enough to not worry about unless / until someone
triggers it and complains.