Fix flaky worklist export test
One of the worklist tests started failing sporadically, because output rows of
the CSV appeared out of the expected order. This is caused from Postgres
generally returning database rows in insertion order -- except when it doesn't.
This is easily fixed by defining a deterministic ordering for all the queries
involved in exports, explicitly. The export selection object now will return
all selected studies, lines, assays, and measurements in a consistent ordering.