Update run_tests.sh script to reuse test database
A significant amount of time in running the EDD test suite is consumed with
creating the database to use for executing tests. Django and `pytest` do this
by finding the configured database connection, creating a new database on the
same server with the database name prefixed with `test_`, then Django runs
through all the migrations for the database. Essentially, make a new empty
database and run `python manage.py migrate`. This work is skipped if the
`--reuse-db` flag is present, AND the database exists.

While it is important to periodically run through the migrations to verify they
still work, we don't need to do this *every* *time*; especially as the
execution time of these migrations accounts for roughly 20% of the total test
execution time. Structuring the change in this way is a compromise. Developers
running `pytest` directly will not get the `--reuse-db` flag by default, like
they would if this change was made in `server/pyproject.toml`. Jenkins builds
will use `run_tests.sh`, and if the build environment has a database, it will
use it. If the Jenkins environment does not have the test database, then there
is no effective change from current behavior.

Additionally, some async tests are still failing when run with the full test
suite. Setting up the run_tests.sh helper script to re-run test failures, so
that there's a chance for load-induced flakiness to be worked around.