Refactor searchers for Select2-based autocompletes
As I was looking into using an autocomplete for import categories, and saw that
doing so would require adding yet another set of imports to the module handling
the Select2 autocompletes, the need to factor out each of the search functions
became clearer. Now, instead of the search module needing to import from all
the items it could potentially search, the items that need searching import
from the search module, and register that they are capable of doing an
autocomplete search.

This refactor consists of adding in a Select2 class to use as a function
decorator, and a SearchRequest class that acts as a facade on the raw HTTP
request for an autocomplete. All the individual search methods are moved out of
the Autocomplete class, and into functions defined next to the relevant Django
model or API for the search.