Remove explicit sleep calls in tests
Some tests included explicit calls to `time.sleep()` in the test, where sleeps
are not strictly necessary to the functioning of the test. Adding in the
`time-machine` library to development dependencies, so these sleeps can be
replaced with a mocked time from Python's C libraries. Now, instead of sleeping
to simulate the passage of time, a test simply moves time to the desired spot
and continues execution without waiting.