Add explicit wait for ICE during startup
Our tests for integration with ICE will sporadically fail when Jenkins launches
the test environment. It appears this happens because ICE finishes startup
before Postgres has a chance to fully initialize, and becomes stuck in an
unhealthy state returning HTTP 404 errors for all requests. Updating the
startup script to explicitly check for a healthy ICE before proceeding,
prevents errors in the integration tests.