Fix link on import button for overview page
Remove the erroneous `as` clause from the URL reverse templatetag, which caused
the true URL for the link to be saved to a variable instead of output in the
page. This made the link the empty string, so clicking the "Import Data" button
would always reload the page. With the URL correctly output to the page, the
button now works again.