Add warning message on export page reload
We currently have support for both GET and POST requests to the export page, so
there can be direct links to an export page or a FORM submit for more complex
export selections. This can cause a problem when reloading the page from a POST
request, as the reload will instead be a GET request with no query parameters.
This change adds in a warning message for when the export selection does not
validate, rather than just displaying the page as-is.